requirements.txt
streamlit
PyPDF2
python-docx
python-pptx
openpyxl
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv